#ci Confiured caching in CI
CircleCi allows to cache build artifacts.
Cargo uses a lock file which is perfect as a cache key, because it
always perfectly reflects the state of the dependencies.
I am not sure about the paths to cache.
The rust docker container has internal environment vars:

- CARGO_INSTALL_ROOT
- CARGO_HOME

But I am unsure how to tell CircleCI about them.