#ci fixing cache paths
the previous experiment seems to have fied the build!
I believe the issue is the intended use case of the rust docker image
and the expectations of CircleCI on docker images do not match.

`tomastomecek/rust` is built to compile project __outside__ the
container, while CircleCI expects to build __inside__ the container.

By setting CARGO_HOME to CIRCLE_WORKING_DIRECTORY, we give the user
inside the container permission to access the cargo crates cache.

These containers are pretty amazing, but I find it impertinent, that
they are being marketed as "simple" or "easy".
In cases like here, they effectivly double (at least) the problem space.